Submission Document
Github repository of the project
 
https://github.com/vbledar/b2b-agreement-contracts.git
https://github.com/vbledar/b2b-agreement-contracts-ui.git
 
Project description
 
The project attempts to tackle b2b agreements based on smart contracts having as a vision to reach a point where a business (eg. hotel) is able to request an agreement from payment providers (mainly financial institutions) in order for the latter to provide payment capabilities to customers wishing to utilise/consume the assets offered by the former. The end goal would to have these agreements performed automatically on a transaction basis where the best offer proposed by a payment provider is selected by the platform.
 
There are two projects involved. The first contains the smart contracts implemented to make it possible for the actual agreement between the entities to be stored on the chain and inherit all the advantages offered by the chain itself, e.g. immutability and a trust less environment (to the extent that this is applicable of course since financial institutions are a highly and extremely regulated environment.
The second part is the implementation of a user interface client which in this stage is mainly used to facilitate a showcase of the smart contract provided functionality making it more tangible for third parties to understand what is going on underneath.
 
Problems that you are solving 
 
The problem being solved derives from the fact that in order to facilitate payments using any current legacy payment method requires an agreement between payment providers and asset owners, but also between consumers of assets/services and payment providers. 
The chain provides the infrastructure by which contracts can be considered immutable and final while they provide transparency of what is actually being implemented since any third party has access to the actual code being executed by the smart contract itself.
An asset owner and a payment provider which try to come in an agreement currently face great costs. Usually it takes a number of meetings to finalise the details which induce costs. Furthermore the achieved agreements are long term and the period in which they extend introduces unpredictable situations because “things” tend to change.

How did you solve them

By utilising smart contracts we can achieve the following:
1. Physical person travel to finalise the agreement is no longer required, thus travelling costs are reduced.
2. Agreements are signed by both parties and can be utilised directly during asset/service booking/consuming.
3. Agreements don’t have to be long term, eventually they can be generated on the fly per transaction in comparison to current hand written agreements which are long term contracts.

What did you learn

It is rather difficult to migrate the legacy processes into blockchain technology. Seems better to think them completely anew, but in a highly regulated environment that may not be really possible.